# [PyTorch Edge] Add Optimized QInt8 Quantize Tensor Arm